+++ src/ReadonlyPlugin.php
@@ -2,17 +2,20 @@
 
 namespace codemonauts\readonly;
 
+use Craft;
 use craft\base\Plugin;
 use craft\events\RegisterComponentTypesEvent;
 use craft\services\Fields;
 use yii\base\Event;
-use codemonauts\readonly\fields\Readonly as ReadonlyField;
-use codemonauts\readonly\feedme\Readonly as ReadonlyFeedme;
+use codemonauts\readonly\fields\ReadonlyField;
+use codemonauts\readonly\feedme\ReadonlyField as ReadonlyFeedme;
 use craft\feedme\events\RegisterFeedMeFieldsEvent;
 use craft\feedme\services\Fields as FeedMeFields;
 
-class Readonly extends Plugin
+class ReadonlyPlugin extends Plugin
 {
+    public string $schemaVersion = '1.0.1';
+
     public function init()
     {
         parent::init();
@@ -22,7 +25,7 @@ class Readonly extends Plugin
         });
 
         // Register field for feed-me plugin if installed
-        if (\Craft::$app->plugins->isPluginEnabled('feed-me')) {
+        if (Craft::$app->plugins->isPluginEnabled('feed-me')) {
             Event::on(FeedMeFields::class, FeedMeFields::EVENT_REGISTER_FEED_ME_FIELDS, function(RegisterFeedMeFieldsEvent $e) {
                     $e->fields[] = ReadonlyFeedme::class;
                 }