+++ src/translations/de/readonly.php
@@ -2,4 +2,6 @@
 
 return [
     'Read-only Field' => 'Schreibgeschütztes Feld',
+    'Template String' => 'Template String',
+    'Twig template string for displaying the value in CP. If not set, the value is printed. You can use {{ value }} and {{ element }}.' => 'Twig Template für die Ausgabe im CP. Wenn nicht gesetzt wird der Wert ausgegeben. Es stehen {{ value }} und {{ element }} zur Verfügung.',
 ];