+++ CHANGELOG.md
@@ -1,13 +1,27 @@
 # Read-only Field Changelog
 
-## Unreleased
+## 2.0.0 - 2022-04-29
+
+### Added
+
+- Craft CMS 4 compatibility
+- php 8.1 compatibility
+- Support for Craft CMS 4 condition builder
+- Twig formatting of value in CP
+- The field is now sortable
+
 ### Changed
-- Fixed background color of icon
+
+- Requires Craft CMS >= 4.0 and php >= 8.0
 
 ## 1.0.1 - 2019-09-14
+
 ### Added
+- 
 - Support for Feed-Me plugin
 
 ## 1.0.0 - 2019-04-06
+
 ### Added
+- 
 - Initial release